fix $(eval foo := $$x) case
in ./repo/android.sh time kati -n, it failed
this was because

repo/android/build/core/definitions.mk
define my-dir
 ..
  $(eval LOCAL_MODULE_MAKEFILE := $$(lastword $$(MAKEFILE_LIST))) \
 ..
endef

this was parsed as
 funcEvalAssign{
   lhs: "LOCAL_MODULE_MAKEFILE"
   op: ":="
   rhs: "$(lastword $(MAKEFILE_LIST))"
 }

so, LOCAL_MODULE_MAKEFILE becomes `$(lastword $(MAKEFILE_LIST))`,
but it should be result of $(lastword $(MAKEFILE_LIST)).

if rhs contains '$' in literal, it should be evaluated again.