[dsan] Split log tracking to two phases
The first phase merges syscall events which didn't finish
soon so that the log has two lines for a single syscall
invocation. Then the next step calculates inputs/outputs.

This makes it easier to track fork/vfork/clone. Before this
change it was tricky to find which process spawned a new
process/thread.