There are 2 features of this change which are important for my Android builds and I know that they are important to some other users. Both aspects are aimed at not regenerating the ninja file needlessly when one is doing an incremental build.
1) Ignore the contents of the TMPDIR environment variable when calculating
the kati stamp:

TMPDIR has to change sometimes between my builds for various complicated
reasons but I think theset also apply to other people who are doing
distributed builds. The result is that the ninja file is regenerated
needlessly. I cannot think of a reason why it *should* be included
therefore it seems reasonable to remove it from consideration.

2) Sort the results of the output from the "find" command.

Filesystems do not all guarantee that 2 successive invocations of find
will return results in the same order. It therefore makes sense to sort
the results.  This patch attempts to sort only if the comparison of 2
file lists fails without it. Therefore the penalty of sorting is only paid
on filesystems which do not return results in the same order every time.
In this way we avoid regenerating the ninja file at times when there is
no relevant reason to do so.