Include implicit outputs in the DepNode graph
The last patch was good enough for simple use cases, but if a dependency
was added to the implicit output, Kati would create a phony rule in
addition to the implicit output.

This patch will cause the rules to be combined into a single DepNode,
combining any variables and dependencies.

Change-Id: I92c9e5b6fdb8bfe49d0cef0a5fa22def4acd6d99