Add deprecated / obsolete variable support
By calling the custom KATI_deprecated_var / KATI_obsolete_var functions,
variables may be marked as deprecated or obsolete.

When accessed or assigned, deprecated variables will print a warning.

When accessed or assigned, obsolete variables will print an error and
stop.

Variables do not need to be set before calling the functions, and will
persist the deprecation warning through a reassignment. This way we can
easily mark variables that are sometimes passed via the environment as
deprecated.

Change-Id: Id04c974c446f471a18cc173f817760f4a02b9239