Fail find parsing for \*
Android hit a few cases where we were searching for *.pyc or
verifiedboot*, and escaping the * as \*. In most cases the entire
argument was just quoted instead.

Kati understands and handles the quotes just fine, but it didn't realize
that \* was special, so it was searching for files that started with a
literal '\'.

We have no good way of freeing allocated memory in the find parser, so
just support stripping a leading \ (\*.pyc works, along with '\('), and
fall back to the system find if another \ is found in a token.