Add some simple sanity checks around PHONY targets
These checks won't apply for all build systems, but for Android, we'd
like to assert:

* That .PHONY targets don't look like real files. This is mostly to
prevent confusion, and just checks to see if any targets marked as phony
have a '/' character in their name.

* That real files don't depend on a PHONY target. If they do, the real
file would be rebuilt on every build execution, potentially causing much
of the build graph to be rebuilt on every execution.

Change-Id: Ibc3023ddccfca5123be41a5124e7bc134e37d2d2