Implement 'final' assignments.
A variable cannot be assigned a new value after the final assignment.
The syntax of the final assignment is
  <variable> =$=<value>
or for the rule-specific variables it is
  <target>: <variable> =$=<value>
The effect is identical to having
  .KATI_READONLY := <variable>
or
  <target>: .KATI_READONLY := variable
that follows the assignment statement.
NOTE that if you still want to use `make`, please avoid having
whitespace after `=$=` (otherwise this whitespace will become
part of the variable's value).

Also, the change rearranges rule parsing and evaluation code.