Add --top_level_phony
This makes the real_to_phony checks assume that any leaf dependencies
that doesn't have a '/' in it is a PHONY target.

We need this in Android since we now have two separate Kati instances
both writing out ninja files from Makefiles. Without this, the
real_to_phony checks don't work when the PHONY target is defined in the
other Kati instance.

Change-Id: Ib0aeee8e57e47b37adef1f5bed871f909cc4af9d