When printing top stats, include count
Instead of just accumulating the time for each instance under a
statistics call, also keep track of how many times it was called:

  *kati*: func shell time: 0.822685 / 148 (114 unique)
  *kati*:   0.019 / 3 cd foo; find -L ../src -name "*.java" -and -not -name ".*"
  ...

Due to line length concerns, I didn't put the average (elapsed/count), but
that's fairly easy to calculate by hand. I also added the '(X unique)'
to understand how unique the calls are.

If it becomes useful in the future, this would be extensible to support
bucketed count -- allowing generation of histograms. I don't expect enough
variation to need that with any stats uses today.

Update to c++17 (the default in Android) to support decomposition
declarations.

Change-Id: I357fb4fdc155dc0be5103f4f3a268f2cdaa4201a