Support '..' in find emulation
This was our biggest source of fallbacks out of the find emulator on
AOSP -- the number of executed shell commands dropped from 53 to 29
after this change. There are only two fallbacks left after this, both
using globs in their finddirs (which turns out to be a much more
complicated change).

I'm seeing similar ratios on our flame (Pixel 4) builds in our internal
trees.

Test: build-aosp_crosshatch.ninja is the same on AOSP master before/after
Test: build-flame.ninja in our internal master is also the same
Change-Id: Ifb0e2e2c08b73726835f832f1461e557ed3c237f