Support '*' in finddirs
To support commands like:

  find foo/*/src -name ...

Add a FindNodes function that fits in between FindDir (which tries to
find a single exact node) and RunFind (which has all of the
find/findleaves/print logic).

This covers the last two cases of `find` in AOSP aosp_flame that weren't
being handled by the find emulator. In the internal flame build there
are 8 (plus 5 more uses of this that are still not handled for other
reasons -- grep/sed, etc)

Test: AOSP build-aosp_flame.ninja is the same before/after
Test: internal build-flame.ninja is the same before/after
Change-Id: I7e9093aee99b1fa244e4aca3383e1efb991887dc