Add new check for non-phony targets that have no commands
There are two related, but separate checks here:

1. --warn_real_no_cmds_or_deps: That an action has no commands or
dependencies that could create the output file. In our ninja output,
these will act similarly to _kati_always_build_ and be considered always
dirty (unless something creates the file before ninja runs, but then
there shouldn't be an action created for it, and we should consider it a
source file).

Actually trying to build these now triggers a ninja error, so this Kati
check is here to catch the issue earlier with better diagnostics.

2. --warn_real_no_cmds: That a rule has no commands, but does have a
dependency that may create the output file. Usually this means that
instead of another action, the dependency should be using
.KATI_IMPLICIT_OUTPUTS to define that it created multiple files.

Under normal circumstances, this shouldn't cause many problems, but can
in some edge cases. I'm also looking to improve our build graph so that
it can function in sandbox-like environments (remote builds, etc). One
of the additions is to remove the output file before we run the commands
that should create it, but this is one of the cases where we don't know
when an output should be created.

Change-Id: Ie330ea49bf3617aa04fa17a3c5aa4dbb6b99b983