CommandEvaluator: Correct Makefile location for multi statement blocks
If a rule has multiple statements to execute, the location information
was not updated accordingly for any but the first statement.

E.g. consider a Makefile with this content:

  | test:
  |    $(warning abc)
  |    $(warning cde)

make would correctly emit:

  Makefile:2 abc
  Makefile:3 cde

Ckati would not update the line number correctly for the second warning,
and instead would emit

  Makefile:2 abc
  Makefile:2 cde

Fix that by updating the evaluator command location from the current
command location.

This fixes testcase/warning.mk that was currenlty failing du to this
mismatch between GNU make and ckati.

Signed-off-by: Matthias Maennich <maennich@google.com>