Bump GNU Make compatibility version
Update the GNU Make compatibility version (MAKE_VERSION) to 4.2.1. The
previously used version 3.81 had been released 2006 the latest 3.x
version is 3.82, released in 2010. With today's toolchains it is quite a
bit of work to still get a 3.81 compiled and odds are that Kati users
are on newer systems anyway (Kati is compiled with C++17).

Lastly, as of today we validate ckati against the GNU Make provided in
Ubuntu 20.04 LTS, which is currently 4.2.1. So, the best we can say
about compatibility is declaring MAKE_VERSION=4.2.1

In order to safeguard this, add explicit steps to use a 4.2.1 GNU Make
for testing in the github actions and add some documentation to achieve
the same result when testing locally.

Signed-off-by: Matthias Maennich <maennich@google.com>