strutil: replace SkipUntil by calls to strcspn
What SkipUntil does is to return the length of the longest string that
does not match any delimiter characters. That is precisely what strcspn
does (except that SkipUntil was treating \0 like any other character).

For the use cases that we have, we can assume that we are not interested
in anything beyond a null terminator of a string. Hence, replace all
uses of SkipUntil with equivalent strcspn invocations.

Signed-off-by: Matthias Maennich <maennich@google.com>