ScopedFrame: remove problematic move constructor
The move constructor was implemented as a copy constructor. While
technically still correct, it creates an unusual interface as the
moved-from object is still perfectly valid. The way the destructor is
implemented, this would have created issues. Luckily, this constructor
was so far never used. Hence, remove it and remove the redundant
deletions of the copy- and assign-constructurs as they are implicitly
deleted if a move constructor is specified.

Signed-off-by: Matthias Maennich <maennich@google.com>