Remove threadpool implementation
The thread pool implementation is currently used for two cases:
1) parallel generation of the ninja output file content
2) second check whether regeneration is necessary

None of those are really on the hot path as rather the whole Eval
machinery is dominating the overall runtime.

There are a couple of issues with the current implementation. The thread
pool always spawns num_jobs threads whenever constructed. Even for the
use case 2) where there are two tasks running in the pool, as many
threads as cores are on the machine are usually spawned, wasting system
resources. In particular, use case 2) can be accomplished with running
the tasks side by side with std::async.

For the use case 1), what we effectively do is map-reduce a massive
string. Each of the tasks produces a substring of the final output and
once they are all generated (in memory), we dump them sorted into the
output file. This makes us build up the output in memory (compute and
memory) and defer the actual writing (i/o) to a later time.

The implementation is changed here to entirely skip using a thread pool.
Instead, we construct a buffered output file stream and pass it around
while we loop over the nodes. That way we can write out the file while
we are still producing the output having and can operate on a single
thread. This also reduces the overall memory foot print.

Those are the measured times from a recent Android aosp/master generator
run:

Without this change:

  (46.543s real, 46.24s user, 7.571s system, 5622MB maxrss)
  (45.176s real, 46.96s user, 5.268s system, 5622MB maxrss)
  (45.118s real, 46.94s user, 5.248s system, 5629MB maxrss)

With this change:

  (45.996s real, 42.05s user, 4.034s system, 4786MB maxrss)
  (45.727s real, 41.69s user, 4.129s system, 4789MB maxrss)
  (45.926s real, 42.19s user, 3.835s system, 4787MB maxrss)

As can be seen, the runtime difference is neglectible, but system
pressure and overall memory consumption significantly improve.

Needless to say, since all use cases can be implemented alternatively,
we can remove the thread pool implementation and it remains less code to
maintain and reason about.

Signed-off-by: Matthias Maennich <maennich@google.com>