Ninja: reduce needless heap allocations
NinjaNodes only ever live in the nodes_ vector and there is therefore no
need to heap allocate them individually as the vector can do that for
us. As a bonus, we do not need to care about the deallocation anymore.

Further, we can now also treat the vector somewhat const and when
iterating over the nodes we can treat them const&, giving some
additional interface clarity.

Signed-off-by: Matthias Maennich <maennich@google.com>