CommandEvaluator: reduce needless allocations and clarify interface
When command vectors are populated by the caller as vector<Command*> and
passed as an input/output parameter via pointer. This interface is
problematic as the contract is not well defined. Clarify that by always
producing the vector within CommandEvaluator::Eval() and return it by
value. Further, the commands are usually only read and never ownership
is transferred. Hence let the vector manage the allocations and don't
bother about that any longer by passing vector<Command> instead.

Signed-off-by: Matthias Maennich <maennich@google.com>