Add KATI_extra_file_deps function
Calling this function with a list of files will add those
files to the kati stamp, causing kati to rerun if those files
are touched.

This is currently not very useful, because the only ways of
observing files will already record kati stamp entries.
(these being $(file) and $(shell))

However, the next commit will add a usecase for it.