Add KATI_shell_no_rerun
We want to add the ability for kati to injest starlark
code to android, but we don't want starlark to rerun
on every invocation of kati. Instead, add a
$(KATI_shell_no_rerun) that will run the shell command
when kati regenerates the ninja file, but not when
checking the stamp file. Then, the starlark files can
be added as dependencies with $(KATI_extra_file_deps).