Fix CI
Ubuntu-22.04 (which is what ubuntu-latest is at the time
of writing) now has clang++-14/clang-format-14 instead of
12. Update those and reformat files. Also, make our github
action based on ubuntu-22.04 explicetly so that when latest
changes we don't break again. (Although in scenarios like
this we would still break)

Also change a forward declaration to an include to resolve
a build failure.