Add an experimental rust port of Kati
Except for unintentional bugs, this should be a drop-in replacement for
ckati, even preserving at least one ckati parsing bug.

This is a relatively literal port of ckati into rust, largely keeping
the same internal APIs and structures, but with some changes. Some of
the larger differences are the replacement of subclasses (with enums or
traits), propagating errors via Result, and using types like
`bytes::Bytes` instead of `std::string`.

Overall performance appears to be fairly similar in Android builds,
though some phases are faster and slower than ckati. There's a lot of
use of `Arc<Mutex<>>` or similar currently, in order to match ckati's
architecture. With refactoring, some of these may be able to go away or
at least reduce to `Rc<RefCell<>>`, as we're not actually multithreaded
except during regen checks.

There are a handful of unsafe uses:
* setting environment variables in `main.rs`, which could be refactored
  to pass the new environment down to the children.
* calling `libc::glob` and `libc::fnmatch` in `fileutil.rs` to preserve
  ckati compatibility.

Change-Id: I478ab27541a1732e1319be73e91f7fd011b46834